requirements.txt
absl-py==0.10.0
astor==0.8.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
-e git+https://github.com/deepcs233/TIN.git@1aacd0c4c30d5e1d334bf023e55b855b59f158db#egg=CUDASHIFT&subdirectory=cuda_shift
cycler==0.10.0
decorator==4.4.2
efficientnet-pytorch==0.7.0
future==0.18.2
gast==0.2.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
MedPy==0.4.0
mock==4.0.2
networkx==2.5
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.1.0.25
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.1.7.4
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
Shapely==1.7.1
SimpleITK==1.2.4
sip==5.4.0
six==1.15.0
sklearn==0.0
tensorboard==2.0.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
xlwt==1.3.0
zipp==3.1.0
